+++ package/Changelog.md
@@ -4,6 +4,14 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+## [1.1.0-pre] - 2021-03-05
+- added experimental code preview feature
+
+## [1.0.1-pre.7] - 2021-03-03
+- moved demystify dll into runtime assembly
+- exposed separator string to add between normal log message and stacktrace
+
 ## [1.0.1-pre.6] - 2021-03-03
 - fix/improve complex syntax highlighting for generic return types
 - prevent syntax highlighting being applied to message